Reorganize header files to an include folder.
