+++ bindings/bindings.cpp
@@ -1,7 +1,9 @@
-#include "bindings.h"
+#include "bindings/bindings.h"
 
 #include <variant>
 #include <vector>
+
+#include "bindings/model/bindings.h"
 
 #define STRINGIFY(x) #x
 #define MACRO_STRINGIFY(x) STRINGIFY(x)
@@ -12,7 +14,12 @@ PYBIND11_MODULE(bindings, m) {
   BindAccelergyInterface(m);
   BindConfigClasses(m);
   BindMappingClasses(m);
-  BindModelClasses(m);
+
+  model_bindings::BindEngine(m);
+  model_bindings::BindLevel(m);
+  model_bindings::BindSparseOptimizationInfo(m);
+  model_bindings::BindTopology(m);
+
   BindProblemClasses(m);
 
 #ifdef VERSION_INFO
