+++ include/bindings/type_casters.h
@@ -5,6 +5,7 @@
 
 // PyBind11 headers
 #include "pybind11/pybind11.h"
+#include "pybind11/stl.h"
 
 // Type casters
 namespace pybind11 {
